msg_types: allow saving caption from command in files
When saving notes, when you reply to a note, the caption of the note itself is used as caption. Allow users to pass the caption as an argument to the bot. For example, saving a file with command "/save test this is a caption" will save "this is a caption" as the file caption.

Signed-off-by: soulr344 <soul@totsuka.gq>